+++ NEWS.md
@@ -1,5 +1,7 @@
 # sora (development version)
 
+* `sora()` has been renamed to `share()`.
+
 # sora 0.0.1
 
 * Initial GitHub release.